refactor: 절대 경로 설정
tsconfig에 주요 경로르를 절대 경로로 추가하여 사용
