+++ src/components/pages/items/BestProducts.tsx
@@ -1,4 +1,4 @@
-import { Product } from '../../../types/product';
+import type { Product } from '@apptypes/product';
 import ProductCard from './ProductCard';
 
 interface BestProductsProps {
