+++ src/components/pages/items/KebabMenu.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { DEFAULT_DELAY_TIME } from '../../../constant';
+import { DEFAULT_DELAY_TIME } from '@constant';
 import './KebabMenu.css';
 
 interface KebabMenuProps {
