+++ src/constant.ts
@@ -13,3 +13,4 @@ export const TOTAL_COUNT = 1000;
 export const DEFAULT_ROW: number = 8;
 export const DEFAULT_BEST_PRODUCT_COUNT: number = 4;
 export const DEFAULT_DELAY_TIME: number = 200;
+export const DEFAULT_PROFILE_IMAGE: string = '/images/icons/ic_mypage.svg';
